% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{angle_between}
\alias{angle_between}
\title{Cacluate angle between vectors}
\usage{
angle_between(a, b, origin = vector3(0, 0, 0))
}
\arguments{
\item{a, b}{Points to measure the angle between}

\item{origin}{(Optional) point to measure the rotation around}
}
\value{
Angle between `a` and `b` in terms of radians
}
\description{
Measure the angle between two points `a` and `b`, with optional argument
`origin` to measure the angle around.
}
