% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motion.R
\name{approx}
\alias{approx}
\title{Linear or constant approximation of vector3}
\usage{
approx(x, y, xout, ...)
}
\arguments{
\item{x}{Indexes for the input values}

\item{y}{The values or vectors to be interpolated between}

\item{xout}{Indexes for the output values}

\item{...}{Additional arguments passed on to stats::approx}
}
\description{
Using `stats::approx`, create a linear or constant approximation
(interpolation) of a value or vector
}
