% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantics-.R, R/semantics-angles.R,
%   R/semantics-axes.R
\docType{data}
\name{common_semantics}
\alias{common_semantics}
\alias{semantics_angles_unity}
\alias{semantics_axes_unity}
\alias{semantics_axes_unreal}
\alias{semantics_axes_opengl}
\title{Predefined axes semantics}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.
}
\usage{
semantics_angles_unity

semantics_axes_unity

semantics_axes_unreal

semantics_axes_opengl
}
\description{
A collection of predefined axis and angle semantics are provided here.
}
\details{
The correctness is, of course, dependent upon whether the semantics are
followed in the virtual world. These are no substitute for visual
inspection.
}
\keyword{datasets}
