% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-look.R, R/layers-restat.R, R/layers.R
\docType{data}
\name{CoordLookAt}
\alias{CoordLookAt}
\alias{StatBin2d3}
\alias{dddr-ggproto}
\alias{StatVector3}
\alias{StatSegment3}
\alias{StatSpoke3}
\title{Spatial Plotting (Prototypes)}
\description{
The plotting in `dddr` is performed using the `ggplot2` extension system. Read more about it at \link[ggplot2]{ggplot2-ggproto}.
}
\keyword{datasets}
