% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dddr-package.R
\docType{package}
\name{dddr-package}
\alias{dddr}
\alias{dddr-package}
\title{dddr: Three-Dimensional Spatial Operations in R}
\description{
A `vctrs`-based package for working with three-diensional
spatial data. It gives three-dimensional objects like
points, velocities, and rotations first-class vector status
in R, enabling usage as column types within `data.frame` and `tibble`.

To learn more about dddr, browse the vignettes:
`browseVignettes(package = "dddr")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MrMallIronmaker/dddr/}
  \item Report bugs at \url{https://github.com/MrMallIronmaker/dddr/issues/}
}

}
\author{
\strong{Maintainer}: Mark Roman Miller \email{mrmillr@stanford.edu}

}
