% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dddr_coords}
\alias{dddr_coords}
\alias{coord_look_at}
\alias{coord_look_at_front}
\alias{coord_look_at_back}
\alias{coord_look_at_top}
\alias{coord_look_at_bottom}
\alias{coord_look_at_left}
\alias{coord_look_at_right}
\title{Spatial Plotting (Coordinates)}
\usage{
coord_look_at(direction, xlim = NULL, ylim = NULL, expand = TRUE, clip = "on")

coord_look_at_front(...)

coord_look_at_back(...)

coord_look_at_top(...)

coord_look_at_bottom(...)

coord_look_at_left(...)

coord_look_at_right(...)
}
\description{
Spatial Plotting (Coordinates)
}
