% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{distance}
\alias{distance}
\title{Determine the distance of a vector or between two vectors}
\usage{
distance(to, from = NULL)
}
\arguments{
\item{to}{Vector to measure the distance of}

\item{from}{Optional. Instead of length being calculated from the origin,
it is calculated from this point instead.}
}
\value{
A numeric vector of distances
}
\description{
Determine the distance of a vector or between two vectors
}
