% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{distance_between}
\alias{distance_between}
\title{Calculate distance between vectors}
\usage{
distance_between(from, to, norm)
}
\arguments{
\item{from, to}{Endpoints of the distance to be calculated}

\item{norm}{optional norm argument from list 'euclidean', 'manhattan', 'L1', 'L2', 'infinity'}
}
\description{
Measure the distance between two points `from` and `to`
}
