% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{euler_angles}
\alias{euler_angles}
\alias{yaw}
\alias{pitch}
\alias{roll}
\title{Euler angles}
\usage{
yaw(q)

pitch(q)

roll(q)
}
\arguments{
\item{q}{Quaternions to extract angles from}
}
\description{
Extract euler angles from a quaternion. These follow Unity's conventions of
yaw around Y, pitch around X, and roll around Z. They are applied in order of
roll-pitch-yaw.
}
