% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extend}
\alias{extend}
\title{extend a point and direction towards a target plane.
 starting at initinal, go in direction until you hit the target plane defined by position and normal vector
 if you want to include vectors traversing backwards, say T, otherwise we get NAs.}
\usage{
extend(
  initial,
  direction,
  target_normal,
  target_point = c(0, 0, 0),
  include_backwards = F
)
}
\description{
extend a point and direction towards a target plane.
 starting at initinal, go in direction until you hit the target plane defined by position and normal vector
 if you want to include vectors traversing backwards, say T, otherwise we get NAs.
}
