% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quat.R
\name{new_quat}
\alias{new_quat}
\title{Internal method for creating a new quat}
\usage{
new_quat(w = double(), x = double(), y = double(), z = double())
}
\arguments{
\item{w, x, y, z}{Quaternion entries, expected to be double vectors}
}
\description{
This follows the new / validate / user-facing pattern recommended in the `vctrs` documentation.
}
\keyword{internal}
