% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{new_vector3}
\alias{new_vector3}
\title{Internal method for creating a new vector3}
\usage{
new_vector3(x = double(), y = double(), z = double())
}
\arguments{
\item{x, y, z}{Vector entries, expected to be double vectors}
}
\description{
This follows the pattern recommended in the `vctrs` documentation.
}
\keyword{internal}
