% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{normalize}
\alias{normalize}
\title{Normalize the length of the vectors to 1 (or a given length)}
\usage{
normalize(v, length = 1)
}
\arguments{
\item{v}{The vectors to normalize}

\item{length}{Optional. All output vectors will have this length.}
}
\description{
All output vectors will have the same direction as the input vectors, but only their lengths will change.
}
