% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{project}
\alias{project}
\title{Vector projection and rejection}
\usage{
project(x, onto)
}
\arguments{
\item{x}{Vectors to project or reject}

\item{onto, from}{The vector to project onto or reject from}
}
\description{
Oftentimes it is useful to constrain a vector to lie on a line or a plane.
Vector projection and rejection is able to do these operations
}
