% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quat.R
\name{quat_helpers}
\alias{quat_helpers}
\alias{upgrade_to_quat}
\alias{ensure_quat}
\title{Quaternion helpers}
\usage{
upgrade_to_quat(v)

ensure_quat(v)
}
\arguments{
\item{v}{a length-4 numeric vector}
}
\description{
Sometimes objects are in a format that is not truly a quaternion but has
unambigous meaning. In those cases, we translate from the length-four
numeric vector to a length-one quaternion
}
\details{
`upgrade` assumes a length 4 numeric, `ensure` checks if it's already a quat first.
}
