% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate.vrm_}
\alias{rotate.vrm_}
\title{rotand = dividend / addend}
\usage{
\method{rotate}{vrm_}(
  rotand,
  rotator = NULL,
  origin = c(0, 0, 0),
  axis = NULL,
  angle = "NULL",
  from = NULL,
  to = NULL,
  ...
)
}
\description{
rotand = dividend / addend
}
