% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{semantics_access}
\alias{semantics_access}
\alias{set_dddr_semantics}
\alias{get_angles_semantics}
\alias{get_axes_semantics}
\alias{get_dddr_semantics}
\title{Semantics access}
\usage{
set_dddr_semantics(axes, angles)

get_angles_semantics()

get_axes_semantics()

get_dddr_semantics()
}
\arguments{
\item{axes, angles}{semantics (see `?semantics_axes` for more details)

  Note that missing values in the `set_dddr_semantics` call are left
  unchanged. Values must be explicitly `NULL` to unset semantics.}
}
\description{
Get and set the global semantics. TODO: more documentation.
}
