% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantics-angles.R
\name{semantics_angles}
\alias{semantics_angles}
\title{Generate angle semantics}
\usage{
semantics_angles(intrinsic = NULL, extrinsic = NULL, hand = NULL)
}
\arguments{
\item{intrinsic, extrinsic}{The intrinsic or extrinsic rotations specified in
order. This can either be a three-letter string like `"ypr"` or a
three-element character vector like `c("yaw", "pitch", "roll")`.}

\item{hand}{`left` or `right` specifying the handedness of the positive angle}
}
\description{
Angle semantics are created using `semantics_angles` functions. As there are
a lot of conventions, we have both an interface to define your own as well as
using ones contributed by the community, such as `semantics_angles_unity` for
the Unity game engine.
}
\seealso{
semantics_angles
}
