% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantics-axes.R
\name{semantics_axes}
\alias{semantics_axes}
\title{Generate axis semantics}
\usage{
semantics_axes(x = NULL, y = NULL, z = NULL, hand = NULL)
}
\arguments{
\item{x, y, z}{String values giving semantic meaning for each positive
direction, relative to the origin. Acceptable values include `up`, `down`,
`left`, `right`, `forward`, and `backward`.}

\item{hand}{`left` or `right` specifying the handedness of the coordinate
system.}
}
\description{
Axis semantics are created using `semantics_axes` functions. As there are a
lot of conventions, we have both an interface to define your own as well as
using ones contributed by the community, such as `semantics_axes_unity` for
the Unity game engine.
}
