% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tait-bryan.R
\name{tait_bryan}
\alias{tait_bryan}
\title{Tait-Bryan angles}
\usage{
tait_bryan(yaw, pitch, roll, unit = c("radians", "degrees"))
}
\arguments{
\item{yaw, pitch, roll}{angles}

\item{unit}{Units of the angles, valid values are `"radians"` and `"degrees"`}
}
\description{
aka euler angles
}
