% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{upgrade_to_vector3}
\alias{upgrade_to_vector3}
\title{Internal helper function for creating a new vector3 from a length-3 numeric vector.}
\usage{
upgrade_to_vector3(v)
}
\arguments{
\item{v}{a length-3 numeric vector}
}
\description{
Internal helper function for creating a new vector3 from a length-3 numeric vector.
}
\keyword{internal}
