% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{vec3_character}
\alias{vec3_character}
\alias{format.dddr_vector3}
\alias{as.character.dddr_vector3}
\alias{as_vector3}
\alias{as_vector3.character}
\title{Vector3 Conversions to and from Character}
\usage{
\method{format}{dddr_vector3}(x, ..., digits = 4)

\method{as.character}{dddr_vector3}(x, ...)

as_vector3(x, ...)

\method{as_vector3}{character}(x, ...)
}
\arguments{
\item{x}{an object to be converted}

\item{...}{arguments passed along to underlying methods `as.character` or
`format`}

\item{digits}{number of decimal digits to produce}
}
\description{
Vector3 Conversions to and from Character
}
