% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{vec_ptype_abbr.vrm_vector3}
\alias{vec_ptype_abbr.vrm_vector3}
\title{A vector3 represennts a 3-dimensional vector from the `vrmvrm` package.}
\usage{
\method{vec_ptype_abbr}{vrm_vector3}(x, ...)
}
\description{
These functions interface with the `vctrs` library
}
