% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{vecjection}
\alias{vecjection}
\alias{project}
\alias{reject}
\title{Vector projection and rejection}
\usage{
project(x, onto)

reject(x, from)
}
\arguments{
\item{x}{Vectors to project or reject}

\item{onto, from}{The vector to project onto or reject from}
}
\description{
Oftentimes it is useful to constrain a vector to lie on a line or a plane.
Vector projection and rejection is able to do these operations. Projection
reduces all vectors to lie on the line of the vector given, and rejection
reduces all vectors to lie on the place normal the the specified vector.
}
