% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{vector3}
\alias{vector3}
\title{Create a 3D vector}
\usage{
vector3(x, y, z)
}
\arguments{
\item{x, y, z}{Numeric vectors representing the the vector's coordinates in the specified dimension.}
}
\description{
Creates a three-dimensional vector given three vectors representing Cartesian coordinates.
}
