% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{vector3_arith}
\alias{vector3_arith}
\alias{vec_arith.dddr_vector3}
\alias{vec_arith.dddr_vector3.default}
\alias{vec_arith.dddr_vector3.dddr_vector3}
\alias{vec_arith.dddr_vector3.numeric}
\alias{vec_arith.dddr_vector3.MISSING}
\alias{vec_arith.numeric.dddr_vector3}
\title{Arithmetic operations on vectors}
\usage{
\method{vec_arith}{dddr_vector3}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{default}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{dddr_vector3}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{numeric}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{MISSING}(op, x, y, ...)

\method{vec_arith.numeric}{dddr_vector3}(op, x, y, ...)
}
\arguments{
\item{op}{Arithmetic operation, i.e, `"+"`, `"-"`, `"*"`, or `"/"`}

\item{x, y}{Operands}

\item{...}{Unused; present for extensibility}
}
\description{
Addition or subtraction requires two `vector3` of equal length or length 1.
In the case one vector is length 1, it is broadcast across the other value.
}
\details{
For multiplication or division, numeric vectors can be length 1, which is
broadcast across all entries, or can be the same length as the vector3
vector, where each element is scaled individually.
}
\examples{
vector3(x = 1:4, y = 2:5, z = 3:6) + vector3(x = 4:7, y = 5:8, z = 6:9)
vector3(x = 1:4, y = 2:5, z = 3:6) - vector3(2, 0, -1)

vector3(x = 1:4, y = 2:5, z = 3:6) * 4
vector3(x = 1:4, y = 2:5, z = 3:6) / c(3, 1, 4, 1)
}
\seealso{
vector3_math
}
