% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{vector3_helpers}
\alias{vector3_helpers}
\alias{upgrade_to_vector3}
\alias{ensure_vector3}
\title{Vector3 helpers}
\usage{
upgrade_to_vector3(v)

ensure_vector3(v)
}
\arguments{
\item{v}{a length-3 numeric vector}
}
\description{
Sometimes objects are in a format that is not truly a vector3 but has
unambigous meaning. In those cases, we translate from the length-three
numeric vector to a length-one vector3.
}
\details{
`upgrade` assumes a length 3 numeric, `ensure` checks if it's already a vector3 first.
}
