% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{vector3_math}
\alias{vector3_math}
\alias{vec_math.vrm_vector3}
\title{Mathematical operations on vectors}
\usage{
\method{vec_math}{vrm_vector3}(.fn, .x, ...)
}
\arguments{
\item{.fn}{Mathematical function to apply}

\item{.x}{Argument to the function}

\item{...}{Unused; reserved for extensibility}
}
\description{
Some summary and cumulant functions can apply to vectors as well. In
particular, `mean`, `sum`, and `cumsum` have meanings when applied to
vector3s.
}
\examples{
mean(vector3(x=1:4, y=2:5, z=3:6))
sum(vector3(x=1:4, y=2:5, z=3:6))
cumsum(vector3(x=1:4, y=2:5, z=3:6))

}
