% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3-operations.R
\name{vector3_prod}
\alias{vector3_prod}
\alias{cross}
\alias{dot}
\title{Vector (cross) and scalar (dot) products}
\usage{
cross(a, b)

dot(a, b)
}
\arguments{
\item{a, b}{Vectors to be multiplied}
}
\description{
Compute the vector and scalar products, also known as the cross and dot
products.
}
\details{
The cross product of two vectors is a vector perpendicular to both inputs and
with the same length as the area of a parallelogram constructed by both
inputs. The dot product returns a scalar value that is the largest (and the
product of the lengths) when the two input vectors are aligned, smallest when
they are 180 degrees apart, and zero when they are perpendicular.
}
