% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{vector3_prop}
\alias{vector3_prop}
\alias{magnitude}
\alias{direction}
\title{Vector properties}
\usage{
magnitude(v)

direction(v)
}
\arguments{
\item{v}{vectors}
}
\description{
Vectors have properties such as magnitude and direction. These are easily
accessible through the `magnitude` and `direction` functions.
}
