% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extend}
\alias{extend}
\title{Extend a point and direction towards a target plane.}
\usage{
extend(
  initial,
  direction,
  target_normal,
  target_point = c(0, 0, 0),
  include_backwards = F
)
}
\arguments{
\item{initial}{Starting point for the ray}

\item{direction}{Direction from which to cast the ray}

\item{target_normal, target_point}{Define the plane that will catch the rays}

\item{include_backwards}{Whether to include rays that travel backwards from `initial` along `direction` to the target plane.}
}
\description{
Starting at `initial`, go in `direction` until you hit the target plane
defined by `target_point` and `target_normal`. If you want to include vectors
going backwards, specify `include_backwards=TRUE`.
}
\details{
This function is intended to be used in conjuction with a plot e.g., to show
where a peron was looking or where an object was pointing over the course of
time or sessions.
}
