% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3.R
\name{field_access}
\alias{field_access}
\alias{$.vrm_vector3}
\title{Vector field access}
\usage{
\method{$}{vrm_vector3}(x, name)
}
\arguments{
\item{x}{Vector3 whose fields to access}

\item{name}{Field name to access; should be one of "x", "y", or "z"}
}
\description{
In order to access fields within each vector, the `$` operator is used.
}
\examples{
vector3(x=1:4, y=2:5, z=3:6)$y
}
