% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quat.R
\name{quat}
\alias{quat}
\title{Create a quaternion}
\usage{
quat(w, x, y, z)
}
\arguments{
\item{w, x, y, z}{Numeric vectors of the quaternion values. It is rare to need
to access these fields.}

\item{axis, angle}{Define a quaternion using axis-angle convention}

\item{from, to}{Define a quaternion using the shortest rotation from `from` to `to`}
}
\description{
Creates a quaternion, a four-dimensional number that represents 3D rotation
well. The values underneath the quaternions are not interpretable for the
naive user; however, one can learn how to read these values. See the vignette
[TODO: link vignette] for an explanation.
}
