% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotation}
\alias{rotation}
\alias{rotate}
\alias{rotate.dddr_vector3}
\alias{rotate.dddr_quat}
\alias{rotate_dddr}
\title{Rotate vectors and quaternions}
\usage{
rotate(rotand, ...)

\method{rotate}{dddr_vector3}(rotand, ...)

\method{rotate}{dddr_quat}(rotand, ...)

rotate_dddr(
  rotand,
  rotator = NULL,
  origin = c(0, 0, 0),
  axis = NULL,
  angle = NULL,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{rotand}{Object to be rotated; can be either a vector or a quaternion}

\item{...}{Additional arguments passed on to underlying S3 methods}

\item{rotator}{(Optional) Quaternion specifying the rotation to perform. If
this argument is not specified, it is constructed using the others.}

\item{origin}{(Optional) The origin of the rotation, i.e, the point that
should not change positions during the rotation.}

\item{axis}{(Optional) The axis of rotation}

\item{angle}{(Optional) The angle of rotation, specified in radians. If this
is not provided, it is calculated using `from` and `to`}

\item{from, to}{(Optional) Instead of specifying an axis-angle pair, or angle
amount, a rotation is performed mapping the direction of `from` to the
direction of `to`.}
}
\description{
Rotate points or quaternions using a handful of ways to define a rotation.
}
\examples{
rotate(vector3(x = 1:4, y = 2:5, z = 3:6), rotator = quat(0, 1, 0, 0))
rotate(vector3(x = 1:4, y = 2:5, z = 3:6), rotator = quat(0, 1, 0, 0), origin = c(1, 2, 3))
rotate(vector3(x = 1:4, y = 2:5, z = 3:6), axis = c(1, 0, 0), angle = pi / 4)
rotate(vector3(x = 1:4, y = 2:5, z = 3:6), from = c(1, 0, 0), to = c(1, 0, 0))
}
