% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3_operations.R
\name{vector3_arith}
\alias{vector3_arith}
\alias{vec_arith.dddr_vector3}
\alias{vec_arith.dddr_vector3.default}
\alias{vec_arith.dddr_vector3.dddr_vector3}
\alias{vec_arith.dddr_vector3.numeric}
\alias{vec_arith.dddr_vector3.MISSING}
\alias{vec_arith.numeric.dddr_vector3}
\title{Arithmetic operations on vectors}
\usage{
\method{vec_arith}{dddr_vector3}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{default}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{dddr_vector3}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{numeric}(op, x, y, ...)

\method{vec_arith.dddr_vector3}{MISSING}(op, x, y, ...)

\method{vec_arith.numeric}{dddr_vector3}(op, x, y, ...)
}
\arguments{
\item{op}{Arithmetic operation, i.e, `"+"`, `"-"`, `"*"`, or `"/"`}

\item{x, y}{Operands}

\item{...}{Unused; present for extensibility}
}
\description{
Vector3s can be added, subtracted, and scaled. This can be done with vectors
of type `vector3` or with numeric vectors of reasonable lengths.
}
\details{
For addition or subtraction, numeric vectors need to be length 3. This vector
is then interpreted as giving the x, y, and z elements, and then is broadcast
across all vector3 entries.

For multiplication or division, numeric vectors can be length 1, which is
broadcast across all entries, or can be the same length as the vector3
vector, where each element is scaled individually.
}
\examples{
vector3(x=1:4, y=2:5, z=3:6) + vector3(x=2, y=0, z=-1)
vector3(x=1:4, y=2:5, z=3:6) + c(2, 0, -1)

vector3(x=1:4, y=2:5, z=3:6) * 4
vector3(x=1:4, y=2:5, z=3:6) * c(3, 1, 4, 1)

}
\seealso{
vector3_math
}
