% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector3-operations.R
\name{vector3_prop}
\alias{vector3_prop}
\alias{magnitude}
\alias{direction}
\title{Vector properties}
\usage{
magnitude(v, norm = c("euclidean", "manhattan", "L1", "L2", "infinity"))

direction(v)
}
\arguments{
\item{v}{vectors}

\item{norm}{optional norm argument from list 'euclidean', 'manhattan', 'L1', 'L2', 'infinity'}
}
\description{
Vectors have properties such as magnitude and direction. These are easily
accessible through the `magnitude` and `direction` functions.
}
